refactor(domain): Lotto 클래스 내 matchCount 메서드 반환 타입 long에서 int로 변경

+++ src/main/java/lotto/domain/Lotto.java
@@ -26,8 +26,8 @@ public class Lotto {
         }
     }
 
-    public long matchCount(Lotto other) {
-        return numbers.stream()
+    public int matchCount(Lotto other) {
+        return (int) numbers.stream()
                 .filter(other.numbers::contains)
                 .count();
     }
